Core/PacketIO: Fixed SMSG_PVP_MATCH_COMPLETE structure
Port From (https://github.com/TrinityCore/TrinityCore/commit/1de5399640261126fed14f4aa0735cc50a4df071)

+++ Source/Game/Networking/Packets/BattleGroundPackets.cs
@@ -532,7 +532,7 @@ namespace Game.Networking.Packets
 
         public override void Write()
         {
-            _worldPacket.WriteUInt8(Winner);
+            _worldPacket.WriteInt32(Winner);
             _worldPacket.WriteInt64(Duration);
             _worldPacket.WriteBit(LogData != null);
             _worldPacket.WriteBits(SoloShuffleStatus, 2);
@@ -542,7 +542,7 @@ namespace Game.Networking.Packets
                 LogData.Write(_worldPacket);
         }
 
-        public byte Winner;
+        public int Winner;
         public long Duration;
         public PVPMatchStatistics LogData;
         public uint SoloShuffleStatus;
